Support unlimited Sequell responses lines
For queries in a private channel (e.g. Discord) or by a player in their
own chat (e.g. WebTiles and Twitch), it's nice to allow Sequell to
return all output lines. This commit adds support for this by checking
should_limit_sequell_lines() in ChatWatcher. Sources for derived classes
can implement this method on their own. The default implementation
limits the response to one line unless the requester is the user who
owns the chat source. For Discord, Cerebot will implement this method to
check that the source is a private channel.